runnable pollers splitting
