+++ src/WorkflowCore.Testing/WorkflowTest.cs
@@ -46,7 +46,7 @@ namespace WorkflowCore.Testing
 
         protected virtual void ConfigureServices(IServiceCollection services)
         {
-            services.AddWorkflow(options => options.UsePollInterval(TimeSpan.FromSeconds(3)));
+            services.AddWorkflow(options => options.UsePollWorkflowsInterval(TimeSpan.FromSeconds(3)));
         }
 
         public string StartWorkflow(TData data)
