+++ test/WorkflowCore.Tests.MongoDB/Scenarios/MongoDelayScenario.cs
@@ -19,7 +19,7 @@ namespace WorkflowCore.Tests.MongoDB.Scenarios
             services.AddWorkflow(cfg =>
             {
                 cfg.UseMongoDB(MongoDockerSetup.ConnectionString, nameof(MongoDelayScenario));
-                cfg.UsePollInterval(TimeSpan.FromSeconds(2));
+                cfg.UsePollWorkflowsInterval(TimeSpan.FromSeconds(2));
             });
         }
     }
