+++ test/WorkflowCore.Tests.SqlServer/Scenarios/SqlServerDelayScenario.cs
@@ -13,7 +13,7 @@ namespace WorkflowCore.Tests.SqlServer.Scenarios
             services.AddWorkflow(cfg =>
             {
                 cfg.UseSqlServer(SqlDockerSetup.ScenarioConnectionString, true, true);
-                cfg.UsePollInterval(TimeSpan.FromSeconds(2));
+                cfg.UsePollWorkflowsInterval(TimeSpan.FromSeconds(2));
             });
         }
     }
